requirements.txt
# This file is here because many Platforms as a Service look for
#	requirements.txt in the root directory of a project.
-r requirements/production.txt
-e git+https://github.com/openpolis/django-popolo.git#egg=django_popolo-dev
unidecode
unicodecsv
django-extensions
ipython
PyYaml